Git/Mercurial: Simplify the sparse checkout pattern
The sparse checkout feature in ORT is an optimization and there a plans
to remove it. It checks out all files recursively residing in the given
VCS path plus all file matching the license filename patterns residing
in any ancestor directory of VCS path.

While this behavior is correct it is also ok to checkout all files
matching any license filename patterns irrespective of the directory.

Do so to allow for further simplifications and to simplify an upcoming
refactoring of the license filename patterns and their handling.

Signed-off-by: Frank Viernau <frank.viernau@here.com>